Shader mode changes now take effect immediately. For ease of testing, made shader mode a read/write property of console.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@3034 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Materials/OOShaderMaterial.m
@@ -385,7 +385,7 @@ static NSString *MacrosToString(NSDictionary *macros);
 	}
 	else
 	{
-		randomSeed = (unsigned)self;
+		randomSeed = (unsigned int)(uintptr_t)self;
 	}
 	savedSeed = RANROTGetFullSeed();
 	ranrot_srand(randomSeed);
